django_002/requirements.txt
alipay==0.7.4
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.13
django-ranged-response==0.2.0
django-rest-swagger==2.2.0
django-simple-captcha==0.5.10
djangorestframework==3.9.1
drf-yasg==1.12.1
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
openapi-codec==1.3.2
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.7
PyYAML==3.13
requests==2.21.0
ruamel.yaml==0.15.86
simplejson==3.16.0
six==1.12.0
tornado==6.0.1
uritemplate==3.0.0
urllib3==1.24.1
